requirements.txt
streamlit
python-dotenv
langchain
langchain-groq
langchain-chroma
huggingface-hub
pandas
langchain-community
sentence-transformers
langchain-huggingface
